chore(aggregated-assets): inject endpoints on the shared DADA api
Addresses @ysitbon's review on #20345: comments 4, 5, 6, 7, 8 and 9.

shared/api-services already exists on develop and its README explicitly names
this migration: 'api-aggregated-assets is a placeholder for DADA ... when it
migrates, its base belongs here as src/services/dada/ rather than as another
standalone createApi'. I had missed it because I checked before rebasing.

Adds services/dada with the endpoint-less base api, then the use-case package
adds its endpoints with injectEndpoints and its cache tag with
enhanceEndpoints({ addTagTypes }), per that README. reducerPath stays the frozen
literal 'assetsDataApi', so one reducer, one middleware and one cache slice
still serve every use case - which the hand-scanning cache selectors depend on.

services/dada carries no extraArgument contract yet, unlike its siblings: DADA
endpoints build absolute URLs and pick prod/staging per request from an
isStaging query arg, so the base query has nothing to own. Migrating that to
extraArgument drops isStaging and touches both apps' store config, so it is
left as a documented TODO rather than smuggled into a relocation.

Layout, per the review:

  index.ts is now export * over public modules only, matching the other three
  domain/api barrels instead of enumerating ~30 named exports.

  internals/ stops being exported and now means what it says. Split by real
  consumers rather than by where things happened to live: market.ts and
  pagination.ts move out (dadaIdToMarketId has 5 app call sites,
  PartialMarketItemResponse 4), while chunkCurrencyIds, deepMergeCryptoAssets,
  emptyAssetsData, collectAllByCategory, assertDadaApiUrl and allSettled have no
  external consumers and stay in.

  collectAllByCategory and emptyAssetsData move to internals with their own
  files; the two public category accessors move to accessors.ts; request
  building and the chunked page fetch move to requests.ts. api.ts is now just
  the endpoint definitions.

The two internal helpers' characterization tests move with them from dada-client
into the api package, unchanged apart from the relative import - they were the
only remaining consumers, so keeping them where they were would have forced the
helpers to stay public. Same 246 tests as before, redistributed: 229 in
dada-client, 17 in the api package.
