+++ domain/api/aggregated-assets/src/market.ts
