--- libs/ledger-live-common/src/dada-client/utils/chunkCurrencyIds.ts
@@ -1,2 +1,0 @@
-export { chunkCurrencyIds } from "@domain/api-aggregated-assets";
-export type { CurrencyIdChunks } from "@domain/api-aggregated-assets";
