--- libs/ledger-live-common/src/dada-client/utils/deepMergeCryptoAssets.ts
@@ -1,1 +1,0 @@
-export { deepMergeCryptoAssets } from "@domain/api-aggregated-assets";
