+++ shared/api-services/src/services/index.ts
@@ -1,5 +1,6 @@
 export * from "./cal";
 export * from "./coinmarketcap";
 export * from "./countervalues";
+export * from "./dada";
 export * from "./push-devices";
 export * from "./swap";
